Updated signal_processing and added unittests
- Changed moving_average since it computed moving average incorrectly.
E.g. the length of the returned array was wrong (off by one). Now it
returns the correct length, but code needs to be changed. Changed
Benchmark3 accordingly. !!! Need to also update the code for the OTFB
!!! Removed outdated unittests for moving_average, that depended on
options which no longer exist.
- Commented out function real_to_cartesian because it is not needed
anywhere, and gives error when maximum is negative
- Added more unittests