OTFB now works with different profiles for beam and LLRF
The OTFB works with voltages etc. computed with a profile of n_llrf many
bins covering one entire turn. The beam profile can be of arbitrary
length and aribtrary number of bins. Question still needs to be
answered, if the LLRF sees the averaged induced voltage of the beam or
the induced voltage of the averaged beam. Need to improve documentation
and clean-up commented parts!