OTFB works with fine and coarse grid
The SPSOneTurnFeedback now works with a fine grid where the beam is, and a coarse grid (1 point per RF-bucket) covering one revolution. The V_fine voltages are used to compute the amplitude and phase corrections used in the tracking of the beam. The V_coarse voltages are used to track the LLRF.