Added unittests for cavity feedback
Replaced test_feedback.py with test_cavity_feedback.py. Unittests now test for a) IQ coordinates of voltage after pre-tracking b) that the RF-voltage seen by the beam after pre-tracking the OTFB is the correct sine-wave c) that the total voltage seen by the beam, including beam-loading generated by the OTFB, is the same as if it were computed with beam-loading from TravellingWaveCavity objects.