Rearrange of omega_rf_d reshape and change of option name fixed_omega_rf
fixed_omega_rf come in place of omega_rf. We should consider that a
change in phi_rf_d should propagate to omega_rf_d in harmonic condition
(unless the machine is very large and the filling time of the cavity
allows for phase changes where the beam is not) and  that changes in
omega_rf_d should propagate in phi_rf_d for fixed frequency condition
(same remark for large machine as above). A suggestion is to incorporate
a bind_omega_and_phase option (in RFStationOptions?)