Update for one-turn feedback and unittests (#120)
* Fixed bug in cavity_feedback.track() and speed-up in modulator()

In SPSOneTurnFeedback.llrf_model(), pi/2 is added to the phase of V_set
so that it is in the Q-channel. This pi/2 must be subtracted from
phi_corr in SPSCavityFeedback.track().
Improved computation time in signal_processing.modulator() by computing
sine and cosine only once.

* Replaced 'slices' by 'profile' in impedance.py

Also added print statement in init of ring.py to show that it's using my
version of the Clean-up branch

* Changed rf_params to rf_station

Changed rf_params to rf_station to better match variable names

* Updated benchmarks to be compadible with latest BLonD clean-up version

This mainly involves a change in the call to the ring constructor.

* Change to moving_average

Changed '!=' comparison to 'is not' to eliminate an error. Also changed
doc-string, since 'center' parameter does no longer exist.

* Removed code that was accidentally copied twice.

* Updated signal_processing and added unittests

- Changed moving_average since it computed moving average incorrectly.
E.g. the length of the returned array was wrong (off by one). Now it
returns the correct length, but code needs to be changed. Changed
Benchmark3 accordingly. !!! Need to also update the code for the OTFB
!!! Removed outdated unittests for moving_average, that depended on
options which no longer exist.
- Commented out function real_to_cartesian because it is not needed
anywhere, and gives error when maximum is negative
- Added more unittests

* Added unittests for wake field and beam loading in test_impulse_response

* Updated cavit_feedback and added unittests

Following the update of moving_average, llrf_model() has been updated to
take again the correct number of points from V_mov_av_prev. Also added
test_feedback.py

* Several changes to cavit_feedback.py

- renamed V_gen to dV_gen in LLRF part (as well as dV_gen_prev,
dV_mov_av_prev); commented out V_gen_mov_av_nest since it is not needed
anywhere
- in track_no_beam, swaped generator_induced_voltage and call to logger,
since the latter depends on V_gen, which is calculated in
generator_induced_voltage
- in track_init, moved computation of V_sum out of the for loop, since
it is only needed after the the initial tracking without beam is
completed
- removed plotting commands in track_init to have less 'clutter'
- updated documentation

* Changed impulse_response

With different bins for the profile and the LLRF, impulse_response now
requires two different times to compute the beam and generator impulse
response. The wake field can be calculated via their own function, since
they are not needed for the remaining computations. Updated unit tests
accordingly.

* FB commissioning works with different number of slices

The feedback now works with h (=4620) slices from 0 to t_rev (=23 micro
s), and the beam can be sliced at arbitrary slices. With the present
version, the commissioning (pre-track) of the OTFB works.
!!! Need to work on beam induced voltage, so that the OTFB also works
with beam loading !!!

* OTFB now works with different profiles for beam and LLRF

The OTFB works with voltages etc. computed with a profile of n_llrf many
bins covering one entire turn. The beam profile can be of arbitrary
length and aribtrary number of bins. Question still needs to be
answered, if the LLRF sees the averaged induced voltage of the beam or
the induced voltage of the averaged beam. Need to improve documentation
and clean-up commented parts!

* Fixed typo in my ring.py

Now prints 'Markus' BLonD'

* Fixed typos in docstrings

Fixed typos for docstrings in impedance_sources and signal_processing

* OTFB works with fine and coarse grid

The SPSOneTurnFeedback now works with a fine grid where the beam is, and a coarse grid (1 point per RF-bucket) covering one revolution. The V_fine voltages are used to compute the amplitude and phase corrections used in the tracking of the beam. The V_coarse voltages are used to track the LLRF.

* Introduced coarse profile in semi-analytic method for less memory useage

To compute the beam-induced voltage on the coarse grid with the semi-analytic method is consumes a lot of memory. To reduce memory consumption, a coarse profile was introduced.

* Changed how V_corr and phi_corr are calculated

While V_ind_beam was correctly calculated in IQ coordinates, the resulting total_voltage in the tracker was different compared to calculating it with TravellingWaveImpedance objects. To correct this:
- the IQ coordinates of V_set are calculated with phase pi/2 - phi_rf
- the real part of V_sum is multiplied by -1
- phi_corr -= -phi_rf + pi/2

* Added unittests for cavity feedback

Replaced test_feedback.py with test_cavity_feedback.py. Unittests now test for a) IQ coordinates of voltage after pre-tracking b) that the RF-voltage seen by the beam after pre-tracking the OTFB is the correct sine-wave c) that the total voltage seen by the beam, including beam-loading generated by the OTFB, is the same as if it were computed with beam-loading from TravellingWaveCavity objects.

* Added option to set gains for each cavity feedback individually

Now individual gains can be set for each SPS cavity feedback by passing a two-element list for either G_llrf or G_tx. If a single  number is provided, both cavity feedback will have the same gain.

* Edited files according to PEP-8

Polished files according to PEP-8 convention

* minor fixes in cavity_feedback

- changed how phi_corr is calculated; now V_sum does no longer need to have its real part change sign
- made cavity_feedback and llrf unittests compatible with latest version of clean-up branch

* Added functions to beam_feedback

- The radial loop computes the mean of all particles, not only those in the profile window. This gets rids of an computationally expansive np.where
- Added SPS_F frequency loop to compare with radial loop results.

* Added unittest for phase loop

Added unittest for the SPS_RL function. It tracks a signle bunch of 1M macro-particles for 200 turn (without radial loop). It computes the difference between the simulated and analytical beam position and compares the result.
Also removed superfluous refence in SPS_F function in beam_feedback.py

* Removed conflicting unittest

The file /general/test_cavity_feedback.py created conflicts during merge because of a different file of the same name in /llrf/test_test_cavity_feedback.py. Moreover, the unit test in /general/test_cavity_feedback.py failed because of obsolete reference values. The file /general/test_cavity_feedback.py was removed from the repository, but the one unit test it contained was moved to /llrf/signal_processing.py as method test_3() of class TestRFCurrent. To avoid further failing unit tests, this test is skipped.

* .travis now installs numpy version >=1.14

Since the cavity_feedback interpolates on complex numbers, it requires numpy version >= 1.14. The change in .travis.yml ensures that.
Also changed test_rf_voltage() in test_cavity_feedback.py to use a rounded result, to avoid failed test due to the 15th digit on different machines.

* Update .travis.yml

Changes in the .travis.yml:
* Build for python 3.5 and 3.6 instead of 2.7 & 3.4
* Caching of miniconda, pip and apt packages to speedup the build process
Build for python 3.5 now succeeds and build 3.6 fails (due to an issue with deepcopy)