Make analysis frequency in units of 1/turn
Was using profile.bin_size to define the frequencies of the coupled bunch
analysis, but that is incorrect.  Ideally, the frequency would be defined
by the revolution period, but that is not constant.  I will try defining
it as inverse numbers of turns, and that should then propagate through
quite nicely, to be seen.