First pass at implementing BeamBaseClass.__iadd__
Made a first pass at implementing magic __iadd__ for in-place addition.
Adds two new functions `add_beam` and `add_particles`, which take
either a BeamBaseClass object or an ArrayLike of new dt and dE values
respectively.

When adding particles, index 0 must be dt, index 1 must be dE.  The new
particles will all be created with _flags set to the active value and
_ids incrementing from n_macroparticles of the original beam.

When adding a Beam, the new particles will keep their `_flags` values,
but have the `_ids` overwritten.  Optionally, a call to
`purge_flagged_entries` can be made when directly calling `add_beam`.
This will not happen when using `__iadd__`.